optimizations regarding data structures. improved error handling code refactoring to follow better modular architecture
